requirements.txt
matplotlib==3.3.2
sentence_transformers==1.1.0
wikipedia==1.4.0
nltk==3.5
numpy==1.19.0
wordcloud==1.8.1
Pillow==8.2.0
scikit_learn==0.24.2
